fix: add beginPath to drawFrame method
Without this you keep adding new subpaths to the same path.
This is a NASTY performance bug, especially in Chrome which starts dropping frames heavily
a few seconds in. It also does not show up in the profiler, because the extra time spent is
registered under "system (self)" rather than the drawFrame execution time, so the only
symptoms are the "long task" warnings with increasing system time, and the shrinking idle time.